+++ controllers/contacts/addContact.js
@@ -1,0 +1,11 @@
+const { Contact } = require("../../models/contact");
+const { ctrlWrapper } = require("../../helpers");
+
+const addContact = async (req, res, __) => {
+  const data = req.body;
+
+  const result = await Contact.create(data);
+  res.status(201).json(result);
+};
+
+module.exports = ctrlWrapper(addContact);
