+++ controllers/contacts/getContactById.js
@@ -1,0 +1,13 @@
+const { Contact } = require("../../models/contact");
+const { ctrlWrapper, HttpError } = require("../../helpers");
+
+const getContactById = async (req, res, __) => {
+  const { contactId } = req.params;
+  const result = await Contact.findById(contactId);
+  if (!result) {
+    throw HttpError(404, "Not found");
+  }
+  res.json(result);
+};
+
+module.exports = ctrlWrapper(getContactById);
