+++ controllers/contacts/updateFavorite.js
@@ -1,0 +1,18 @@
+const { Contact } = require("../../models/contact");
+const { HttpError, ctrlWrapper } = require("../../helpers");
+
+const updateFavorite = async (req, res, __) => {
+  const data = req.body;
+
+  const { contactId } = req.params;
+  const result = await Contact.findByIdAndUpdate(contactId, data, {
+    new: true,
+  });
+  if (!result) {
+    throw HttpError(404, "Not found. Contact with such id doesn't exist");
+  }
+
+  res.json(result);
+};
+
+module.exports = ctrlWrapper(updateFavorite);
